improvement: split service employee
