+++ apps/api/employee/src/app/app.controller.ts
@@ -2,15 +2,15 @@ import { Controller } from "@nestjs/common";
 
 import { AppService } from "./app.service";
 import { MessagePattern } from "@nestjs/microservices";
-import { TEmployeePaginationArgs } from "@uninus/entities";
+import { ISelectRequest, TEmployeePaginationArgs } from "@uninus/entities";
 
 @Controller()
 export class AppController {
   constructor(private readonly appService: AppService) {}
 
   @MessagePattern("get_employees")
-  getEmployees({ type, where, orderBy, page, perPage }: TEmployeePaginationArgs) {
-    return this.appService.getEmployees({ type, where, orderBy, page, perPage });
+  getEmployees(payload: TEmployeePaginationArgs) {
+    return this.appService.getEmployees(payload);
   }
 
   @MessagePattern("get_total_employees")
@@ -19,12 +19,17 @@ export class AppController {
   }
 
   @MessagePattern("get_lecturer")
-  getLecturer(id: string) {
-    return this.appService.getLecturer(id);
+  getLecturer(payload: { id: string }) {
+    return this.appService.getLecturer(payload);
   }
 
   @MessagePattern("get_academic_staff")
-  getAcademicStaff(id: string) {
-    return this.appService.getAcademicStaff(id);
+  getAcademicStaff(payload: { id: string }) {
+    return this.appService.getAcademicStaff(payload);
+  }
+
+  @MessagePattern("get_employee_categories")
+  async getCategories(payload: ISelectRequest) {
+    return await this.appService.getCategories(payload);
   }
 }
