+++ apps/api/general/src/app/app.service.ts
@@ -61,7 +61,6 @@ import {
   TGeneralResponse,
   TCreateScholarshipRequest,
   TRegistrationPathResponse,
-  TEmployeeCategoriesResponse,
   TQuestionResponse,
   TTotalRegistransRes,
 } from "@uninus/entities";
@@ -1938,25 +1937,5 @@ export class AppService {
     return {
       message: `Berhasil menghapus beasiswa `,
     };
-  }
-
-  async getCategories({ search, id }: ISelectRequest): Promise<TEmployeeCategoriesResponse> {
-    const employeeCategories = await this.prisma.employeeCategories.findMany({
-      where: {
-        id: id && Number(id),
-        name: {
-          ...(search && { contains: search }),
-          mode: "insensitive",
-        },
-      },
-      select: {
-        id: true,
-        name: true,
-      },
-    });
-    if (!employeeCategories) {
-      throw new RpcException(new NotFoundException("Data tidak ditemukan"));
-    }
-    return employeeCategories;
   }
 }
