+++ libs/api/services/src/index.ts
@@ -3,3 +3,4 @@ export * from "./student";
 export * from "./prisma";
 export * from "./user";
 export * from "./finance";
+export * from "./employee";
