--- libs/entities/src/types/filter/index.ts
@@ -1,5 +1,0 @@
-export type TRpcException = {
-  statusCode: number;
-  message: string;
-  error: string;
-};
